add: home doesn't show groupFinder when confirmed
